@@ -12,24 +12,26 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
+    "@angular/animations": "7.2.14",
+    "@angular/common": "7.2.14",
+    "@angular/compiler": "7.2.14",
+    "@angular/core": "7.2.14",
+    "@angular/forms": "7.2.14",
+    "@angular/http": "7.2.14",
+    "@angular/platform-browser": "7.2.14",
+    "@angular/platform-browser-dynamic": "7.2.14",
+    "@angular/router": "7.2.14",
     "bootstrap": "^4.3.1",
     "core-js": "^2.4.1",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19"
+    "rxjs": "^6.5.1",
+    "tslib": "^1.9.0",
+    "zone.js": "^0.8.29"
   },
   "devDependencies": {
-    "@angular/cli": "~1.7.0",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular/cli": "7.3.8",
+    "@angular/compiler-cli": "7.2.14",
+    "@angular/language-service": "7.2.14",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
@@ -44,6 +46,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "3.2.4"
   }
 }
